[clientapitest] Move unit test operator== away from API header
XYZ operator== is used only in unit test code. Due to taking also
timestamps into account it is "samples are exactly the same" rather
than "sample values are similar enough" type of test - which is
unlikely to be useful outside unit testing context.

Move the operator from public API header to where it is needed.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>