core.pro: use QT6_INSTALL_LIBDIR when provided
When using meta-qt6 on Yocto, the QT_INSTALL_LIBS points to the host folder
for Qt6 libs install.

On device we end up with sensorfwd looking in a directory like this:
/media/LuneOS/testing/webos-ports/tmp-glibc/work/pinephone-webos-linux/sensorfw/0.14.4+gitAUTOINC+24ab29a7b7-r0/recipe-sysroot/usr/lib/sensord-qt6

To avoid that, make it possible to use an alternative qmake variable
named QT6_INSTALL_LIBDIR when its value is defined by the build toolchain.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>