[hybrisadapter] Implement doubleStopReader quirk. JB#62669
In C2 disabling gyroscope when also accelerometer is enabled can lead
to accelerometer reporting to freeze. Sensor drivers are not available
for analysis / fixing, but it seems that forcing gyroscope restart in
different datarate resumes accelerometer reporting.

As this is a workaround for a device specific problem, this behavior
is disabled by default and should be enabled on as-needed basis in
device specific sensorfwd configuration.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>